--- cognite/powerops/client/_generated/affr_bid/__init__.py
@@ -1,3 +1,0 @@
-from cognite.powerops.client._generated.affr_bid._api_client import AFRRBidAPI
-
-__all__ = ["AFRRBidAPI"]
