+++ cognite/powerops/client/_generated/afrr_bid/__init__.py
@@ -1,0 +1,3 @@
+from cognite.powerops.client._generated.afrr_bid._api_client import AFRRBidAPI
+
+__all__ = ["AFRRBidAPI"]
