+++ cognite/powerops/client/_generated/afrr_bid/_api/bid_document_query.py
@@ -5,22 +5,21 @@ from typing import TYPE_CHECKING
 
 from cognite.client import data_modeling as dm, CogniteClient
 
-from cognite.powerops.client._generated.affr_bid.data_classes import (
+from cognite.powerops.client._generated.afrr_bid.data_classes import (
     DomainModelApply,
-    BidRow,
-    BidRowApply,
-    BidRow,
-    BidRowApply,
-    BidMethod,
-    BidMethodApply,
+    BidDocument,
+    BidDocumentApply,
+    PriceArea,
+    PriceAreaApply,
 )
 from ._core import DEFAULT_QUERY_LIMIT, QueryBuilder, QueryStep, QueryAPI, T_DomainModelList, _create_edge_filter
 
 if TYPE_CHECKING:
     from .alert_query import AlertQueryAPI
+    from .bid_row_query import BidRowQueryAPI
 
 
-class BidRowQueryAPI(QueryAPI[T_DomainModelList]):
+class BidDocumentQueryAPI(QueryAPI[T_DomainModelList]):
     def __init__(
         self,
         client: CogniteClient,
@@ -33,13 +32,13 @@ class BidRowQueryAPI(QueryAPI[T_DomainModelList]):
 
         self._builder.append(
             QueryStep(
-                name=self._builder.next_name("bid_row"),
+                name=self._builder.next_name("bid_document"),
                 expression=dm.query.NodeResultSetExpression(
                     from_=self._builder[-1].name if self._builder else None,
                     filter=filter_,
                 ),
-                select=dm.query.Select([dm.query.SourceSelector(self._view_by_write_class[BidRowApply], ["*"])]),
-                result_cls=BidRow,
+                select=dm.query.Select([dm.query.SourceSelector(self._view_by_write_class[BidDocumentApply], ["*"])]),
+                result_cls=BidDocument,
                 max_retrieve_limit=limit,
             )
         )
@@ -49,18 +48,16 @@ class BidRowQueryAPI(QueryAPI[T_DomainModelList]):
         external_id_prefix: str | None = None,
         space: str | list[str] | None = None,
         limit: int | None = DEFAULT_QUERY_LIMIT,
-        retrieve_linked_bid: bool = False,
-        retrieve_method: bool = False,
+        retrieve_price_area: bool = False,
     ) -> AlertQueryAPI[T_DomainModelList]:
-        """Query along the alert edges of the bid row.
+        """Query along the alert edges of the bid document.
 
         Args:
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
             limit: Maximum number of alert edges to return. Defaults to 25. Set to -1, float("inf") or None
                 to return all items.
-            retrieve_linked_bid: Whether to retrieve the linked bid for each bid row or not.
-            retrieve_method: Whether to retrieve the method for each bid row or not.
+            retrieve_price_area: Whether to retrieve the price area for each bid document or not.
 
         Returns:
             AlertQueryAPI: The query API for the alert.
@@ -85,64 +82,84 @@ class BidRowQueryAPI(QueryAPI[T_DomainModelList]):
                 max_retrieve_limit=limit,
             )
         )
-        if retrieve_linked_bid:
-            self._query_append_linked_bid(from_)
-        if retrieve_method:
-            self._query_append_method(from_)
+        if retrieve_price_area:
+            self._query_append_price_area(from_)
         return AlertQueryAPI(self._client, self._builder, self._view_by_write_class, None, limit)
+
+    def bids(
+        self,
+        external_id_prefix: str | None = None,
+        space: str | list[str] | None = None,
+        limit: int | None = DEFAULT_QUERY_LIMIT,
+        retrieve_price_area: bool = False,
+    ) -> BidRowQueryAPI[T_DomainModelList]:
+        """Query along the bid edges of the bid document.
+
+        Args:
+            external_id_prefix: The prefix of the external ID to filter on.
+            space: The space to filter on.
+            limit: Maximum number of bid edges to return. Defaults to 25. Set to -1, float("inf") or None
+                to return all items.
+            retrieve_price_area: Whether to retrieve the price area for each bid document or not.
+
+        Returns:
+            BidRowQueryAPI: The query API for the bid row.
+        """
+        from .bid_row_query import BidRowQueryAPI
+
+        from_ = self._builder[-1].name
+
+        edge_filter = _create_edge_filter(
+            dm.DirectRelationReference("power-ops-types", "PartialBid"),
+            external_id_prefix=external_id_prefix,
+            space=space,
+        )
+        self._builder.append(
+            QueryStep(
+                name=self._builder.next_name("bids"),
+                expression=dm.query.EdgeResultSetExpression(
+                    filter=edge_filter,
+                    from_=from_,
+                ),
+                select=dm.query.Select(),
+                max_retrieve_limit=limit,
+            )
+        )
+        if retrieve_price_area:
+            self._query_append_price_area(from_)
+        return BidRowQueryAPI(self._client, self._builder, self._view_by_write_class, None, limit)
 
     def query(
         self,
-        retrieve_linked_bid: bool = False,
-        retrieve_method: bool = False,
+        retrieve_price_area: bool = False,
     ) -> T_DomainModelList:
         """Execute query and return the result.
 
         Args:
-            retrieve_linked_bid: Whether to retrieve the linked bid for each bid row or not.
-            retrieve_method: Whether to retrieve the method for each bid row or not.
+            retrieve_price_area: Whether to retrieve the price area for each bid document or not.
 
         Returns:
             The list of the source nodes of the query.
 
         """
         from_ = self._builder[-1].name
-        if retrieve_linked_bid:
-            self._query_append_linked_bid(from_)
-        if retrieve_method:
-            self._query_append_method(from_)
+        if retrieve_price_area:
+            self._query_append_price_area(from_)
         return self._query()
 
-    def _query_append_linked_bid(self, from_: str) -> None:
-        view_id = self._view_by_write_class[BidRowApply]
+    def _query_append_price_area(self, from_: str) -> None:
+        view_id = self._view_by_write_class[PriceAreaApply]
         self._builder.append(
             QueryStep(
-                name=self._builder.next_name("linked_bid"),
+                name=self._builder.next_name("price_area"),
                 expression=dm.query.NodeResultSetExpression(
                     filter=dm.filters.HasData(views=[view_id]),
                     from_=from_,
-                    through=self._view_by_write_class[BidRowApply].as_property_ref("linked_bid"),
+                    through=self._view_by_write_class[BidDocumentApply].as_property_ref("price_area"),
                     direction="outwards",
                 ),
                 select=dm.query.Select([dm.query.SourceSelector(view_id, ["*"])]),
                 max_retrieve_limit=-1,
-                result_cls=BidRow,
-            ),
-        )
-
-    def _query_append_method(self, from_: str) -> None:
-        view_id = self._view_by_write_class[BidMethodApply]
-        self._builder.append(
-            QueryStep(
-                name=self._builder.next_name("method"),
-                expression=dm.query.NodeResultSetExpression(
-                    filter=dm.filters.HasData(views=[view_id]),
-                    from_=from_,
-                    through=self._view_by_write_class[BidRowApply].as_property_ref("method"),
-                    direction="outwards",
-                ),
-                select=dm.query.Select([dm.query.SourceSelector(view_id, ["*"])]),
-                max_retrieve_limit=-1,
-                result_cls=BidMethod,
+                result_cls=PriceArea,
             ),
         )
