+++ cognite/powerops/client/_generated/afrr_bid/_api/price_area_total_capacity_allocation_down.py
@@ -9,7 +9,7 @@ from cognite.client import CogniteClient
 from cognite.client import data_modeling as dm
 from cognite.client.data_classes import Datapoints, DatapointsArrayList, DatapointsList, TimeSeriesList
 from cognite.client.data_classes.datapoints import Aggregate
-from cognite.powerops.client._generated.affr_bid.data_classes._price_area import _create_price_area_filter
+from cognite.powerops.client._generated.afrr_bid.data_classes._price_area import _create_price_area_filter
 from ._core import DEFAULT_LIMIT_READ, INSTANCE_QUERY_LIMIT
 
 ColumnNames = Literal[
@@ -26,7 +26,7 @@ ColumnNames = Literal[
 ]
 
 
-class PriceAreaCapacityPriceUpQuery:
+class PriceAreaTotalCapacityAllocationDownQuery:
     def __init__(
         self,
         client: CogniteClient,
@@ -51,7 +51,7 @@ class PriceAreaCapacityPriceUpQuery:
         limit: int | None = None,
         include_outside_points: bool = False,
     ) -> DatapointsList:
-        """`Retrieve datapoints for the `price_area.capacity_price_up` timeseries.
+        """`Retrieve datapoints for the `price_area.total_capacity_allocation_down` timeseries.
 
         **Performance guide**:
             In order to retrieve millions of datapoints as efficiently as possible, here are a few guidelines:
@@ -76,11 +76,11 @@ class PriceAreaCapacityPriceUpQuery:
         Examples:
 
             In this example,
-            we are using the time-ago format to get raw data for the 'my_capacity_price_up' from 2 weeks ago up until now::
+            we are using the time-ago format to get raw data for the 'my_total_capacity_allocation_down' from 2 weeks ago up until now::
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> client = AFRRBidAPI()
-                >>> price_area_datapoints = client.price_area.capacity_price_up(external_id="my_capacity_price_up").retrieve(start="2w-ago")
+                >>> price_area_datapoints = client.price_area.total_capacity_allocation_down(external_id="my_total_capacity_allocation_down").retrieve(start="2w-ago")
         """
         external_ids = self._retrieve_timeseries_external_ids_with_extra()
         if external_ids:
@@ -110,7 +110,7 @@ class PriceAreaCapacityPriceUpQuery:
         limit: int | None = None,
         include_outside_points: bool = False,
     ) -> DatapointsArrayList:
-        """`Retrieve numpy arrays for the `price_area.capacity_price_up` timeseries.
+        """`Retrieve numpy arrays for the `price_area.total_capacity_allocation_down` timeseries.
 
         **Performance guide**:
             In order to retrieve millions of datapoints as efficiently as possible, here are a few guidelines:
@@ -135,11 +135,11 @@ class PriceAreaCapacityPriceUpQuery:
         Examples:
 
             In this example,
-            we are using the time-ago format to get raw data for the 'my_capacity_price_up' from 2 weeks ago up until now::
+            we are using the time-ago format to get raw data for the 'my_total_capacity_allocation_down' from 2 weeks ago up until now::
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> client = AFRRBidAPI()
-                >>> price_area_datapoints = client.price_area.capacity_price_up(external_id="my_capacity_price_up").retrieve_array(start="2w-ago")
+                >>> price_area_datapoints = client.price_area.total_capacity_allocation_down(external_id="my_total_capacity_allocation_down").retrieve_array(start="2w-ago")
         """
         external_ids = self._retrieve_timeseries_external_ids_with_extra()
         if external_ids:
@@ -171,9 +171,9 @@ class PriceAreaCapacityPriceUpQuery:
         uniform_index: bool = False,
         include_aggregate_name: bool = True,
         include_granularity_name: bool = False,
-        column_names: ColumnNames | list[ColumnNames] = "capacityPriceUp",
+        column_names: ColumnNames | list[ColumnNames] = "totalCapacityAllocationDown",
     ) -> pd.DataFrame:
-        """`Retrieve DataFrames for the `price_area.capacity_price_up` timeseries.
+        """`Retrieve DataFrames for the `price_area.total_capacity_allocation_down` timeseries.
 
         **Performance guide**:
             In order to retrieve millions of datapoints as efficiently as possible, here are a few guidelines:
@@ -194,7 +194,7 @@ class PriceAreaCapacityPriceUpQuery:
             uniform_index: If only querying aggregates AND a single granularity is used, AND no limit is used, specifying `uniform_index=True` will return a dataframe with an equidistant datetime index from the earliest `start` to the latest `end` (missing values will be NaNs). If these requirements are not met, a ValueError is raised. Default: False
             include_aggregate_name: Include 'aggregate' in the column name, e.g. `my-ts|average`. Ignored for raw time series. Default: True
             include_granularity_name: Include 'granularity' in the column name, e.g. `my-ts|12h`. Added after 'aggregate' when present. Ignored for raw time series. Default: False
-            column_names: Which property to use for column names. Defauts to capacityPriceUp
+            column_names: Which property to use for column names. Defauts to totalCapacityAllocationDown
 
 
         Returns:
@@ -203,11 +203,11 @@ class PriceAreaCapacityPriceUpQuery:
         Examples:
 
             In this example,
-            we are using the time-ago format to get raw data for the 'my_capacity_price_up' from 2 weeks ago up until now::
+            we are using the time-ago format to get raw data for the 'my_total_capacity_allocation_down' from 2 weeks ago up until now::
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> client = AFRRBidAPI()
-                >>> price_area_datapoints = client.price_area.capacity_price_up(external_id="my_capacity_price_up").retrieve_dataframe(start="2w-ago")
+                >>> price_area_datapoints = client.price_area.total_capacity_allocation_down(external_id="my_total_capacity_allocation_down").retrieve_dataframe(start="2w-ago")
         """
         external_ids = self._retrieve_timeseries_external_ids_with_extra(column_names)
         if external_ids:
@@ -248,9 +248,9 @@ class PriceAreaCapacityPriceUpQuery:
         uniform_index: bool = False,
         include_aggregate_name: bool = True,
         include_granularity_name: bool = False,
-        column_names: ColumnNames | list[ColumnNames] = "capacityPriceUp",
+        column_names: ColumnNames | list[ColumnNames] = "totalCapacityAllocationDown",
     ) -> pd.DataFrame:
-        """Retrieve DataFrames for the `price_area.capacity_price_up` timeseries in Timezone.
+        """Retrieve DataFrames for the `price_area.total_capacity_allocation_down` timeseries in Timezone.
 
         **Performance guide**:
             In order to retrieve millions of datapoints as efficiently as possible, here are a few guidelines:
@@ -271,7 +271,7 @@ class PriceAreaCapacityPriceUpQuery:
             uniform_index: If only querying aggregates AND a single granularity is used, AND no limit is used, specifying `uniform_index=True` will return a dataframe with an equidistant datetime index from the earliest `start` to the latest `end` (missing values will be NaNs). If these requirements are not met, a ValueError is raised. Default: False
             include_aggregate_name: Include 'aggregate' in the column name, e.g. `my-ts|average`. Ignored for raw time series. Default: True
             include_granularity_name: Include 'granularity' in the column name, e.g. `my-ts|12h`. Added after 'aggregate' when present. Ignored for raw time series. Default: False
-            column_names: Which property to use for column names. Defauts to capacityPriceUp
+            column_names: Which property to use for column names. Defauts to totalCapacityAllocationDown
 
 
         Returns:
@@ -280,13 +280,13 @@ class PriceAreaCapacityPriceUpQuery:
         Examples:
 
             In this example,
-            get weekly aggregates for the 'my_capacity_price_up' for the first month of 2023 in Oslo time:
+            get weekly aggregates for the 'my_total_capacity_allocation_down' for the first month of 2023 in Oslo time:
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> from datetime import datetime, timezone
                 >>> client = AFRRBidAPI()
-                >>> price_area_datapoints = client.price_area.capacity_price_up(
-                ...     external_id="my_capacity_price_up").retrieve_dataframe_in_timezone(
+                >>> price_area_datapoints = client.price_area.total_capacity_allocation_down(
+                ...     external_id="my_total_capacity_allocation_down").retrieve_dataframe_in_timezone(
                 ...         datetime(2023, 1, 1, tzinfo=ZoneInfo("Europe/Oslo")),
                 ...         datetime(2023, 1, 2, tzinfo=ZoneInfo("Europe/Oslo")),
                 ...         aggregates="average",
@@ -332,9 +332,9 @@ class PriceAreaCapacityPriceUpQuery:
             return None
 
     def _retrieve_timeseries_external_ids_with_extra(
-        self, extra_properties: ColumnNames | list[ColumnNames] = "capacityPriceUp"
+        self, extra_properties: ColumnNames | list[ColumnNames] = "totalCapacityAllocationDown"
     ) -> dict[str, list[str]]:
-        return _retrieve_timeseries_external_ids_with_extra_capacity_price_up(
+        return _retrieve_timeseries_external_ids_with_extra_total_capacity_allocation_down(
             self._client,
             self._view_id,
             self._filter,
@@ -350,7 +350,7 @@ class PriceAreaCapacityPriceUpQuery:
         include_aggregate_name: bool,
         include_granularity_name: bool,
     ) -> pd.DataFrame:
-        if isinstance(column_names, str) and column_names == "capacityPriceUp":
+        if isinstance(column_names, str) and column_names == "totalCapacityAllocationDown":
             return df
         splits = sum(included for included in [include_aggregate_name, include_granularity_name])
         if splits == 0:
@@ -363,7 +363,7 @@ class PriceAreaCapacityPriceUpQuery:
         return df
 
 
-class PriceAreaCapacityPriceUpAPI:
+class PriceAreaTotalCapacityAllocationDownAPI:
     def __init__(self, client: CogniteClient, view_id: dm.ViewId):
         self._client = client
         self._view_id = view_id
@@ -376,8 +376,8 @@ class PriceAreaCapacityPriceUpAPI:
         space: str | list[str] | None = None,
         limit: int = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
-    ) -> PriceAreaCapacityPriceUpQuery:
-        """Query timeseries `price_area.capacity_price_up`
+    ) -> PriceAreaTotalCapacityAllocationDownQuery:
+        """Query timeseries `price_area.total_capacity_allocation_down`
 
         Args:
             name: The name to filter on.
@@ -388,16 +388,16 @@ class PriceAreaCapacityPriceUpAPI:
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
-            A query object that can be used to retrieve datapoins for the price_area.capacity_price_up timeseries
+            A query object that can be used to retrieve datapoins for the price_area.total_capacity_allocation_down timeseries
             selected in this method.
 
         Examples:
 
-            Retrieve all data for 5 price_area.capacity_price_up timeseries:
+            Retrieve all data for 5 price_area.total_capacity_allocation_down timeseries:
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> client = AFRRBidAPI()
-                >>> price_areas = client.price_area.capacity_price_up(limit=5).retrieve()
+                >>> price_areas = client.price_area.total_capacity_allocation_down(limit=5).retrieve()
 
         """
         filter_ = _create_price_area_filter(
@@ -409,7 +409,7 @@ class PriceAreaCapacityPriceUpAPI:
             filter,
         )
 
-        return PriceAreaCapacityPriceUpQuery(
+        return PriceAreaTotalCapacityAllocationDownQuery(
             client=self._client,
             view_id=self._view_id,
             timeseries_limit=limit,
@@ -425,7 +425,7 @@ class PriceAreaCapacityPriceUpAPI:
         limit: int = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
     ) -> TimeSeriesList:
-        """List timeseries `price_area.capacity_price_up`
+        """List timeseries `price_area.total_capacity_allocation_down`
 
         Args:
             name: The name to filter on.
@@ -436,15 +436,15 @@ class PriceAreaCapacityPriceUpAPI:
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
-            List of Timeseries price_area.capacity_price_up.
+            List of Timeseries price_area.total_capacity_allocation_down.
 
         Examples:
 
-            List price_area.capacity_price_up and limit to 5:
+            List price_area.total_capacity_allocation_down and limit to 5:
 
-                >>> from cognite.powerops.client._generated.affr_bid import AFRRBidAPI
+                >>> from cognite.powerops.client._generated.afrr_bid import AFRRBidAPI
                 >>> client = AFRRBidAPI()
-                >>> price_areas = client.price_area.capacity_price_up.list(limit=5)
+                >>> price_areas = client.price_area.total_capacity_allocation_down.list(limit=5)
 
         """
         filter_ = _create_price_area_filter(
@@ -455,7 +455,7 @@ class PriceAreaCapacityPriceUpAPI:
             space,
             filter,
         )
-        external_ids = _retrieve_timeseries_external_ids_with_extra_capacity_price_up(
+        external_ids = _retrieve_timeseries_external_ids_with_extra_total_capacity_allocation_down(
             self._client, self._view_id, filter_, limit
         )
         if external_ids:
@@ -464,21 +464,21 @@ class PriceAreaCapacityPriceUpAPI:
             return TimeSeriesList([])
 
 
-def _retrieve_timeseries_external_ids_with_extra_capacity_price_up(
+def _retrieve_timeseries_external_ids_with_extra_total_capacity_allocation_down(
     client: CogniteClient,
     view_id: dm.ViewId,
     filter_: dm.Filter | None,
     limit: int,
-    extra_properties: ColumnNames | list[ColumnNames] = "capacityPriceUp",
+    extra_properties: ColumnNames | list[ColumnNames] = "totalCapacityAllocationDown",
 ) -> dict[str, list[str]]:
     limit = float("inf") if limit is None or limit == -1 else limit
-    properties = ["capacityPriceUp"]
-    if extra_properties == "capacityPriceUp":
+    properties = ["totalCapacityAllocationDown"]
+    if extra_properties == "totalCapacityAllocationDown":
         ...
-    elif isinstance(extra_properties, str) and extra_properties != "capacityPriceUp":
+    elif isinstance(extra_properties, str) and extra_properties != "totalCapacityAllocationDown":
         properties.append(extra_properties)
     elif isinstance(extra_properties, list):
-        properties.extend([prop for prop in extra_properties if prop != "capacityPriceUp"])
+        properties.extend([prop for prop in extra_properties if prop != "totalCapacityAllocationDown"])
     else:
         raise ValueError(f"Invalid value for extra_properties: {extra_properties}")
 
@@ -508,7 +508,9 @@ def _retrieve_timeseries_external_ids_with_extra_capacity_price_up(
         )
         result = client.data_modeling.instances.query(query)
         batch_external_ids = {
-            node.properties[view_id]["capacityPriceUp"]: [node.properties[view_id].get(prop, "") for prop in extra_list]
+            node.properties[view_id]["totalCapacityAllocationDown"]: [
+                node.properties[view_id].get(prop, "") for prop in extra_list
+            ]
             for node in result.data["nodes"].data
         }
         total_retrieved += len(batch_external_ids)
