+++ cognite/powerops/client/_generated/afrr_bid/_api_client.py
