+++ cognite/powerops/client/_generated/afrr_bid/data_classes/__init__.py
