+++ cognite/powerops/client/_generated/afrr_bid/data_classes/_bid_method.py
