+++ cognite/powerops/client/_generated/afrr_bid/data_classes/_core.py
